requirements.txt
asgiref==3.10.0
boto3==1.40.65
botocore==1.40.65
Django==5.2.7
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
jmespath==1.0.1
mysql-connector-python==9.5.0
openpyxl==3.1.5
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
